requirements.txt
tensorflow-gpu==1.12
keras<=2.2.4
pillow
sklearn
matplotlib
numpy